move base classes into module folder to ease installation
Installing classes outside of packages makes setup.py usage
complicated and clutters the root directory.
Move Base classes into modules directory
